Update FrskyLog2Gpx.ps1
Added -ForceUTC option to parse logfile time as UTC instead of local time.  The ayvri.com site always shows UTC time even though it says LOCAL TIME so we just write local time and mark it as UTC in the output kml.  This is the wrong time but it forces ayvri to display local time.